pkg/api/core: move pkg/api/ref
