+++ pkg/apis/core/helper/ref/BUILD
@@ -10,7 +10,7 @@ go_test(
     name = "go_default_test",
     srcs = ["ref_test.go"],
     embed = [":go_default_library"],
-    importpath = "k8s.io/kubernetes/pkg/api/ref",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/ref",
     deps = [
         "//pkg/api/legacyscheme:go_default_library",
         "//pkg/apis/core:go_default_library",
@@ -23,7 +23,7 @@ go_test(
 go_library(
     name = "go_default_library",
     srcs = ["ref.go"],
-    importpath = "k8s.io/kubernetes/pkg/api/ref",
+    importpath = "k8s.io/kubernetes/pkg/apis/core/helper/ref",
     deps = [
         "//pkg/apis/core:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/meta:go_default_library",
