+++ pkg/client/clientset_generated/internalclientset/typed/core/internalversion/event_expansion.go
@@ -24,8 +24,8 @@ import (
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/kubernetes/pkg/api/ref"
 	api "k8s.io/kubernetes/pkg/apis/core"
+	"k8s.io/kubernetes/pkg/apis/core/helper/ref"
 	k8s_api_v1 "k8s.io/kubernetes/pkg/apis/core/v1"
 )
