+++ pkg/kubectl/.import-restrictions
@@ -6,7 +6,7 @@
                 "k8s.io/kubernetes/pkg/api/events",
                 "k8s.io/kubernetes/pkg/api/legacyscheme",
                 "k8s.io/kubernetes/pkg/apis/core/helper/pod",
-                "k8s.io/kubernetes/pkg/api/ref",
+                "k8s.io/kubernetes/pkg/apis/core/helper/ref",
                 "k8s.io/kubernetes/pkg/apis/core/helper/resource",
                 "k8s.io/kubernetes/pkg/apis/core/helper/service",
                 "k8s.io/kubernetes/pkg/apis/core/v1/helper/pod",
